requirements.txt
streamlit
openfoodfacts
zxing-cpp
Pillow
# streamlit-barcode-scanner @ git+https://github.com/FloWide/streamlit-barcode-scanner.git
